use styles.ts
