+++ task_manager/src/apps/landing.tsx
@@ -1,64 +1,15 @@
 import React, { useEffect, useState } from 'react';
 import { Button, Caption1, Card, CardHeader, CardPreview, Checkbox, makeStyles, Text, Title1, tokens } from '@fluentui/react-components';
 import image from '../assets/image.png';
-import { Navigate, useNavigate } from 'react-router';
+import { useNavigate } from 'react-router';
+import { landingStyles } from '../styles';
 
 interface LandingProps {
   tasks: string[];
 }
 
-const flex = {
-  gap: "16px",
-  display: "flex",
-};
-
-const useStyles = makeStyles({
-  parent: {
-    display: "flex",
-    justifyContent: "center",
-    alignItems: "center",
-  },
-  main: {
-    ...flex,
-    flexDirection: "column",
-    justifyContent: "center",
-    alignItems: "center",
-  },
-
-  row: {
-    ...flex,
-    flexWrap: "wrap",
-    justifyContent: "center",
-    alignItems: "center",
-    width: "60%"
-  },
-
-  card: {
-    width: "200px",
-    maxWidth: "100%",
-    height: "fit-content",
-  },
-
-  caption: {
-    color: tokens.colorNeutralForeground3,
-  },
-
-  smallRadius: { borderRadius: tokens.borderRadiusSmall },
-
-  grayBackground: {
-    backgroundColor: tokens.colorNeutralBackground3,
-  },
-  onboardButton: {
-    position: 'fixed',
-    top: '10px',
-    right: '10px',
-    zIndex: 1000,
-    paddingTop: "10px"
-  }
-});
-
 const Landing = ({tasks}: LandingProps) => {
-  const styles = useStyles();
+  const styles = landingStyles();
   const items = tasks.map((task) => { return { name: task, description: "test_description", selected: false } });
   const [selectedTasks, setSelectedTasks] = useState(items);
   const [numSelected, setNumSelected] = useState(0);
